Stream Enrich: support AT_TIMESTAMP as initial position (closes #3360)
